Continuing fiat balance modularization
